+++ security/cmd/node_agent_k8s/workloadapi/workloadapi.go
@@ -22,7 +22,7 @@ import (
 
 	rpc "istio.io/gogo-genproto/googleapis/google/rpc"
 
-	"istio.io/istio/security/cmd/node_agent_k8s/binder"
+	"istio.io/istio/security/pkg/flexvolume/binder"
 	pb "istio.io/istio/security/proto"
 )
